Refactor CurrencyRateService - a file split into a new folder + get rid of useless methods
